Refactor class hierarchy
This should solve the issues I've been having with
virtual inheritance and the way methods are overriden.
