+++ src/d3d9/d3d9_device.h
@@ -14,8 +14,6 @@ namespace dxvk {
   public:
     virtual ~D3D9Device();
 
-    HRESULT STDMETHODCALLTYPE QueryInterface(REFIID riid, void** ppvObject) final override;
-
     HRESULT STDMETHODCALLTYPE TestCooperativeLevel() final override {
       Logger::err(str::format(__func__, " stub"));
       throw DxvkError("Not supported");
@@ -38,6 +36,12 @@ namespace dxvk {
     }
 
   protected:
-    virtual ID3D11Device* device() const = 0;
+    // This constructor should be called first in the class hierarchy.
+    // It initializes the associated D3D11 device.
+    D3D9Device(D3D9Adapter& adapter, HWND hFocusWindow, D3DPRESENT_PARAMETERS& pp);
+
+    D3D9Adapter& m_adapter;
+    Com<ID3D11Device> m_device;
+    Com<ID3D11DeviceContext> m_ctx;
   };
 }
