+++ src/d3d9/d3d9_query.h
@@ -4,7 +4,7 @@
 
 namespace dxvk {
   /// Implements query creation.
-  class D3D9DeviceQuery: public virtual ComObject<D3D9Device> {
+  class D3D9DeviceQuery: public virtual D3D9Device {
     HRESULT STDMETHODCALLTYPE CreateQuery(D3DQUERYTYPE Type,
       IDirect3DQuery9** ppQuery) final override;
   };
