+++ src/d3d9/d3d9_shader.h
@@ -1,10 +1,11 @@
 #pragma once
 
 #include "d3d9_device.h"
+#include "d3d9_device_child.h"
 
 namespace dxvk {
   /// Vertex shader device functions.
-  class D3D9DeviceVertexShader: public virtual ComObject<D3D9Device> {
+  class D3D9DeviceVertexShader: public virtual D3D9Device {
     HRESULT STDMETHODCALLTYPE CreateVertexShader(const DWORD* pFunction,
       IDirect3DVertexShader9** ppShader) final override {
       Logger::err(str::format(__func__, " stub"));
@@ -61,7 +62,7 @@ namespace dxvk {
   };
 
   /// Pixel shader device functions.
-  class D3D9DevicePixelShader: public virtual ComObject<D3D9Device> {
+  class D3D9DevicePixelShader: public virtual D3D9Device {
     HRESULT STDMETHODCALLTYPE CreatePixelShader(const DWORD* pFunction,
       IDirect3DPixelShader9** ppShader) final override {
       Logger::err(str::format(__func__, " stub"));
@@ -116,4 +117,19 @@ namespace dxvk {
       throw DxvkError("Not supported");
     }
   };
+
+  /// Shader class.
+  template <typename Base, typename I>
+  class D3D9Shader final: public ComObject<Base>, D3D9DeviceChild {
+  public:
+    HRESULT STDMETHODCALLTYPE GetDevice(IDirect3DDevice9** ppDevice) override;
+
+    HRESULT STDMETHODCALLTYPE GetFunction(void* pFunction, UINT* pSizeOfData) override;
+  private:
+    std::vector<DWORD> m_func;
+    Com<I> m_shader;
+  };
+
+  using D3D9VertexShader = D3D9Shader<IDirect3DVertexShader9, ID3D11VertexShader>;
+  using D3D9PixelShader = D3D9Shader<IDirect3DPixelShader9, ID3D11PixelShader>;
 }
