+++ src/d3d9/d3d9_viewport.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "d3d9_include.h"
+#include "d3d9_device.h"
 
 namespace dxvk {
   /// Viewport-related functions implementation.
-  class D3D9DeviceViewport: public virtual IDirect3DDevice9 {
+  class D3D9DeviceViewport: public virtual D3D9Device {
   public:
     HRESULT STDMETHODCALLTYPE GetViewport(D3DVIEWPORT9* pViewport) final override;
